pushed plant items in garden into react components
