+++ client/package.json
@@ -7,6 +7,7 @@
     "axios": "^0.18.0",
     "classnames": "^2.2.6",
     "jwt-decode": "^2.2.0",
+    "nodemailer": "^6.4.14",
     "react": "^16.14.0",
     "react-debounce-input": "^3.2.2",
     "react-dom": "^16.6.3",
