+++ client/src/components/layout/Garden.js
@@ -1,24 +1,26 @@
-import { decodeBase64 } from "bcryptjs";
 import React, { useEffect, useState } from "react";
 import store from "../../store";
+import GardenItem from "./GardenItem";
 
 function Garden() {
   const [garden, setGarden] = useState([]);
   const user = store.getState().auth.user.id;
   useEffect(() => {
-    fetch("api/users/" + user, { method: "GET" }).then((user) => {
-      user.json().then((user) => {
-        console.log(user.garden);
+    fetch("api/users/" + user, { method: "GET" })
+    .then((user) => {
+      user.json()
+      .then((user) => {
         setGarden(user.garden);
+        // console.log(garden);
       });
     });
   });
+  // console.log(garden);
   return (
     <div>
     <h3>My Garden</h3>
       {garden.map((plant) => (
-        <p>{plant.name}</p>
-      ))}
+        <GardenItem {...plant} key={plant.id} />      ))}
     </div>
   );
 }
